Task 1
Created a simple chatbot that uses python as a base, accessing queries trough a .json file. The program uses fuzzywuzzy to be approximately match what the user mentions in their question to what is stored in prompts.json which contains the tuples. Coded to answer a variety of general questions through command line.